V4 R&D Rounds 8-13: 90 techniques + bug fixes + cleanup
R&D Rounds 8-13 implementing 90 inference/training techniques from
latest 2026 research, plus codebase-wide audit and bug fixes.

Round 8: FA4 Blackwell backend, SF-NorMuon/AMUSE/MONA optimizers,
KARA/MomentKV/KVpop/CONF-KV eviction methods.

Round 9: P-EAGLE + Lookahead Gate speculative decoding, Pruned BPE +
ToaST tokenization, Jet-Long/RoPE-ID/LeRoPE/LaMPE positional encoding.

Round 10: FastServe/Libra/FASER/Kairos schedulers, curriculum learning,
training-time augmentation, SYNPRO synthetic data, SeeDNorm/DyT/Keel
normalization.

Round 11: Unified Radix Cache (HiCache), Elbow/Alloc-MoE/LDA MoE
optimization, SPPO/PS-PPO/EVPO/GRPO-OR advanced RL algorithms.

Round 12: AdaMX/SharQ/MosaicQuant adaptive quantization, AoH data-free
sparse attention, RetMask retrieval head optimization, offline top-K
logits + chunked KL + sequence truncation + prefix OPD distillation.

Round 13: General upkeep - fixed 8 bugs (4 critical: kvpop n_kv missing,
conf_kv positions attr, moe_optim total_budgets typo, aoh_retmask weight
layout), added missing __init__.py, wired flex_decoding, verified 342
files compile + 356/361 tests pass.

Also: cleanup of dead code (old distillation data, serving server,
recursive self-play, research_team, etc.), .devin scratchpad and
profiling scripts.

+++ research/sampling_utils.py
@@ -59,7 +59,7 @@ def top_p_filter_logits(logits: torch.Tensor, top_p: float,
 
     # Step 4: Map removal back to topk indices, then to original vocab indices.
     sorted_remove = torch.zeros_like(sorted_probs)
-    sorted_remove.scatter_(-1, sorted_idx, remove.float())
+    sorted_remove.scatter_(-1, sorted_idx, remove.to(sorted_remove.dtype))
     # Map back to topk space
     topk_remove = torch.zeros_like(topk_probs)
     topk_remove.scatter_(-1, sorted_idx, sorted_remove)
